requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
pillow>=10.0.0
reportlab>=4.0.0
xhtml2pdf>=0.2.13
python-docx>=1.1.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# 2FA Support
pyotp>=2.9.0
qrcode>=7.4.2

# Email
django-anymail>=10.2

# Throttling and Security
django-ratelimit>=4.1.0

# Payments
requests>=2.32.0
cryptography>=42.0.0

# Assistant LLM provider
openai>=1.61.0

